Прикручена БД на сервере. Прикручено логирование на сервере. Прикручена авторизация.

+++ server/src/main/java/alexrm84/myFirstCloudBox/server/DistributorHandler.java
@@ -1,5 +1,6 @@
 package alexrm84.myFirstCloudBox.server;
 
+import alexrm84.myFirstCloudBox.common.Command;
 import alexrm84.myFirstCloudBox.common.FileMessage;
 import alexrm84.myFirstCloudBox.common.SystemMessage;
 import io.netty.channel.ChannelHandlerContext;
@@ -7,34 +8,44 @@ import io.netty.channel.ChannelInboundHandlerAdapter;
 import io.netty.util.ReferenceCountUtil;
 
 public class DistributorHandler extends ChannelInboundHandlerAdapter {
+    private boolean authorization = false;
 
     @Override
     public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
         Worker worker = new Worker(ctx);
         try {
-            if (msg == null){
+            if (msg == null) {
                 return;
             }
-            if (msg instanceof SystemMessage){
-                SystemMessage systemMessage = (SystemMessage)msg;
-                switch (systemMessage.getTypeMessage()){
-                    case Refresh:
-                        ctx.writeAndFlush(systemMessage.setPathsList(worker.refreshFiles(systemMessage.getRequestedPath())));
-                        break;
-                    case CheckPath:
-                        worker.checkPath(ctx, systemMessage);
-                        break;
-                    case ReceiveFiles:
-                        worker.sendFiles(ctx, systemMessage);
-                        break;
-                    case DeleteFiles:
-                        worker.deleteFiles(ctx, systemMessage);
-                        break;
+            if (authorization) {
+                if (msg instanceof SystemMessage){
+                    SystemMessage systemMessage = (SystemMessage)msg;
+                    switch (systemMessage.getTypeMessage()){
+                        case Refresh:
+                            ctx.writeAndFlush(systemMessage.setPathsList(worker.refreshFiles(systemMessage.getRequestedPath())));
+                            break;
+                        case CheckPath:
+                            worker.checkPath(ctx, systemMessage);
+                            break;
+                        case ReceiveFiles:
+                            worker.sendFiles(ctx, systemMessage);
+                            break;
+                        case DeleteFiles:
+                            worker.deleteFiles(ctx, systemMessage);
+                            break;
+                    }
                 }
-            }
-            if (msg instanceof FileMessage){
-                FileMessage fileMessage = (FileMessage)msg;
-                worker.writeFile(ctx, fileMessage);
+                if (msg instanceof FileMessage){
+                    FileMessage fileMessage = (FileMessage)msg;
+                    worker.writeFile(ctx, fileMessage);
+                }
+            }else {
+                if (msg instanceof SystemMessage){
+                    SystemMessage systemMessage = (SystemMessage) msg;
+                    if (systemMessage.getTypeMessage().equals(Command.Authorization)) {
+                        authorization = worker.authorization(ctx, systemMessage);
+                    }
+                }
             }
         }finally {
             ReferenceCountUtil.release(msg);
